requirements.txt
streamlit
openai
pandas
python-dotenv
matplotlib
numpy
faiss-cpu
